airflow/my-etl-pipleine/requirements.txt
apache-airflow
apache-airflow-providers-snowflake
apache-airflow-providers-postgres
pandas
snowflake-connector-python
python-dotenv